+++ src/flags.ts
@@ -230,11 +230,12 @@ export const optDefs: commandLineUsage.OptionDefinition[] = [
     defaultValue: defaults.traceCategories.join(','),
   },
   {
-    name: 'collate',
+    name: 'partition',
     description:
-        `Whether to collate result table by measurement (when multiple ` +
-        `measurements are in use)`,
-    type: Boolean,
+      `What to partition the results table by. Use "measurement" ` +
+      `when multiple measurements are in use and you want multiple ` +
+      `smaller results tables.`,
+    type: String,
   },
 ];
 
@@ -266,7 +267,7 @@ export interface Opts {
   trace: boolean;
   'trace-log-dir': string;
   'trace-cat': string;
-  collate: boolean;
+  partition: string;
 
   // Extra arguments not associated with a flag are put here. These are our
   // benchmark names/URLs.